backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
djangorestframework==3.14.0
drf-yasg==1.21.4
flake8==5.0.4
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==21.3
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytz==2022.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
